Move LLMError from railtracks.exceptions into the llm package
LLMError now lives in railtracks.llm.errors and inherits from RTLLMError,
eliminating the dependency the llm package had on railtracks.exceptions.
Also replaces a NodeInvocationError raised inside _litellm_wrapper with
ModelError to fix the same cross-package violation. All built_nodes,
related tests, and exception tests updated to import from the new location.

Closes #1096
