+++ packages/railtracks/src/railtracks/exceptions/__init__.py
@@ -2,7 +2,6 @@ from .errors import (
     ContextError,
     FatalError,
     GlobalTimeOutError,
-    LLMError,
     NodeCreationError,
     NodeInvocationError,
 )
@@ -13,7 +12,6 @@ __all__ = [
     "NodeCreationError",
     "NodeInvocationError",
     "GlobalTimeOutError",
-    "LLMError",
     "ContextError",
     "VisualExtraRequiredError",
 ]
